Move light constants to separate module
