+++ homeassistant/components/light/device_trigger.py
@@ -10,7 +10,7 @@ from homeassistant.core import CALLBACK_TYPE, HomeAssistant
 from homeassistant.helpers.trigger import TriggerActionType, TriggerInfo
 from homeassistant.helpers.typing import ConfigType
 
-from . import DOMAIN
+from .const import DOMAIN
 
 TRIGGER_SCHEMA = vol.All(
     toggle_entity.TRIGGER_SCHEMA,
